requirements.txt
celery==5.4.0
Django==5.0.4
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0